Gracefully tolerate numerical instability
When probabilities are incorporated into a ProbabilisitcSubcircuitBlock,
they are now enforced to:

 - be >= 0,
 - be <= 1, and
 - sum to 1.

The deviation from these requirements is tracked, and---if larger than
CUTOFF_WARN, an error is show.  Execution is halted if it is also
greater than CUTOFF_FAIL.

A KEEP_RAW parameter is added to give experts access to the
unnormalized probabilities.

This replaces existing pyGSTi-specific machinery that validated the
output of pyGSTi model objects.

Signed-off-by: Antonio Russo <arusso@sandia.gov>