Do a "full_reload" by default
Unlike a regular reload, in which the namespace of the reloaded module
is rewritten with the new module's properties [which does NOT remove
existing attributes of the module], a "full_reload" returns a new
reference entirely.

Existing references to the old module *remain*, but are not accessible
through the newly-loaded module.

Signed-off-by: Antonio Russo <arusso@sandia.gov>