Cleanup idle_gate into idle_gates
Implement idle_gate as a generator, allowing it to more easily return 0
idle gates, when appropriate.  This allows more logic to move into the
function, and for uniform calls to it.

The structure of the pyGSTi circuit changes, however, so the tests must
be updated.  At the same time, we make them slightly more robust.

Signed-off-by: Antonio Russo <arusso@sandia.gov>